+++ src/tab-logic.js
@@ -1,7 +1,13 @@
-const siteContent = document.getElementById('site-content')
+import { addHomeContent } from "./home-main.js"
+// const { addHomeContent } = require("./home-main")
 
-const menuButton = document.getElementById('menu')
+const siteContent = document.getElementById('site-content')
+const homeContent = document.getElementById('home-content')
+
+
+
 const homeButton = document.getElementById('home')
+const menuButton = document.getElementById('menu')
 const aboutUsButton = document.getElementById('about-us')
 let currentChild = homeContent
 
@@ -10,11 +16,18 @@ function tabSwitch(btn) {
         console.log('CLICK')
         siteContent.removeChild(currentChild)
         currentChild = btn
+        if (currentChild == homeButton) {
+            addHomeContent();
+            console.log('addhomecontent');
+            console.log(currentChild);
+        }
         siteContent.appendChild(currentChild)
     })
 }
 
-tabSwitch(menuButton)
+
+
 tabSwitch(homeButton)
+tabSwitch(menuButton)
 tabSwitch(aboutUsButton)
